Various tweaks
- added additional filters and actions to vue Element
- Broke Element component into a wrapper so we can modify the rendering component from filters
- Allow actions/triggers to have unlimited number of arguments
