+++ packages/preview/src/index.ts
@@ -10,11 +10,13 @@ const appContext = editorApp._context
 // Components
 import { InlineEditor } from './components/InlineEditor'
 import Element from './components/Element.vue'
+import ElementWrapper from './components/ElementWrapper.vue'
 import SortableContent from './components/SortableContent.vue'
 import RenderValue from './components/RenderValue.vue'
 import ElementIcon from './components/ElementIcon.vue'
 
 editorApp.component('Element', Element)
+editorApp.component('ElementWrapper', ElementWrapper)
 editorApp.component('InlineEditor', InlineEditor)
 editorApp.component('SortableContent', SortableContent)
 editorApp.component('RenderValue', RenderValue)
